minor: improved event page loading time

+++ src/server/service/participant.service.ts
@@ -51,6 +51,12 @@ export async function create(eventId: string, userId: string) {
     logger.info("ParticipantsService.create");
     CommonValidator.INPUT("Event Id", eventId);
     CommonValidator.INPUT("User Id", userId);
+    const isAlreadyParticipant = await isParticipant(eventId);
+
+    if (isAlreadyParticipant.data) {
+      return IResponse.toJSON(400, "Participant already exists", null);
+    }
+
     const response = await ParticipantDAO.create(eventId, userId);
     return IResponse.toJSON(201, "Participant created", response);
   } catch (error) {
@@ -69,6 +75,12 @@ export async function remove(eventId: string, userId: string) {
     logger.info("ParticipantsService.remove");
     CommonValidator.INPUT("Event Id", eventId);
     CommonValidator.INPUT("User Id", userId);
+
+    const isAlreadyParticipant = await isParticipant(eventId);
+    if (!isAlreadyParticipant.data) {
+      return IResponse.toJSON(400, "Participant does not exist", null);
+    }
+
     const response = await ParticipantDAO.remove(eventId, userId);
     return IResponse.toJSON(204, "Participant removed", response);
   } catch (error) {
